+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   with (allows date/time parsing, etc.).
 - Drop support for all Ruby versions prior to 2.3 (would be easy to
   re-introduce, but I imagine anyone using this gem is building a new app).
+- Rename Spot::API to Spot::Client
 
 ## v0.0.1, 10 July 2016